refactor: 依存関係の整理をしてみた
